added multiple versions of the implementation
There are 4 different attempts at this problem with this data set. the first is a simple 2 layer zero-shot classification that checks labels then sub-labels under a single label until a best match is found. the second version uses a single layer and checks all the sub-labels at equal value. the third takes the top 3 sub-labels and returns them. the fourth one does a much larger check of each individual set of data in the json and returns the top score(this has the best output but very bad run time).